added some instructions

+++ source/chip8.cpp
@@ -7,7 +7,7 @@
 
 
 constexpr uint16_t START_ADDRESS{0x200};
-constexpr unsigned int FONTSET_START_ADDRESS {0x50};
+constexpr uint16_t FONTSET_START_ADDRESS {0x50};
 static constexpr unsigned int FONTSET_SIZE{80};
